backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
firebase-admin==6.5.0
uuid==1.30
PyJWT==2.8.0